+++ app/src/main/java/com/example/soundnest_android/data/SongRepository.kt
@@ -4,6 +4,6 @@ import com.example.soundnest_android.ui.comments.Comment
 import com.example.soundnest_android.ui.songs.Song
 
 interface SongRepository {
-    suspend fun getSongById(songId: Long): Song
-    suspend fun getCommentsForSong(songId: Long): List<Comment>
+    suspend fun getSongById(songId: Int): Song
+    suspend fun getCommentsForSong(songId: Int): List<Comment>
 }
