+++ app/src/main/java/com/example/soundnest_android/ui/songs/Song.kt
@@ -3,7 +3,7 @@ package com.example.soundnest_android.ui.songs
 import java.io.Serializable
 
 data class Song(
-    val id: Long,             // ← nuevo
+    val id: Int,
     val title: String,
     val artist: String,
     val coverResId: Int
